Restructure project according to Domain Driven Design.
