+++ src/modules/dictionary/index.ts
@@ -1,0 +1,2 @@
+export * from './DictionaryConvert';
+export * from './DictionaryType';
