+++ src/modules/index.ts
@@ -1,0 +1,5 @@
+import {Dictionary} from './dictionary';
+
+export namespace Modules {
+    export const dictionary = Dictionary;
+}
