+++ src/modules/privateAccess/PrivateAccessConvert.ts
@@ -1,0 +1,11 @@
+import { IPrivateAccessData } from ".";
+
+export function convertPrivateAccess(
+  privateAccessToConvert: any
+): IPrivateAccessData {
+  return {
+    tableName: privateAccessToConvert.table_name,
+    recordId: privateAccessToConvert.record_id,
+    recordUuid: privateAccessToConvert.record_uuid
+  };
+}
