+++ src/modules/privateAccess/index.ts
@@ -1,0 +1,2 @@
+export * from './PrivateAccessConvert';
+export * from './PrivateAccessType';
