+++ src/modules/process/ProcessType.d.ts
@@ -1,0 +1,15 @@
+import { IValueData } from "@/modules/core";
+import { IReportOutputData } from "@/modules/report";
+
+export interface IProcessLogData {
+  uuid: string;
+  instanceUuid: string;
+  isError: boolean;
+  summary: string;
+  resultTableName: string;
+  isProcessing: boolean;
+  lastRun: number;
+  parametersList?: any;
+  parameters?: IValueData[];
+  output: IReportOutputData;
+}
