+++ src/modules/window/WindowType.d.ts
