--- src/utils/ADempiere/apiConverts/persistence/types.d.ts
@@ -1,18 +1,0 @@
-interface IEntityData {
-  id: number;
-  uuid: string;
-  tableName: string;
-  attributes: any;
-}
-
-interface IEntityListData {
-  nextPageToken: string;
-  recordCount: number;
-  recordsList: IEntityData[];
-}
-
-interface ITranslationData {
-  language: string;
-  uuid: string;
-  values: any;
-}
