--- src/utils/ADempiere/apiConverts/privateAccess/index.ts
@@ -1,11 +1,0 @@
-import { IPrivateAccessData } from "./types";
-
-export function convertPrivateAccess(
-  privateAccessToConvert: any
-): IPrivateAccessData {
-  return {
-    tableName: privateAccessToConvert.table_name,
-    recordId: privateAccessToConvert.record_id,
-    recordUuid: privateAccessToConvert.record_uuid
-  };
-}
